Uses the easy peasy values directly instead from the store
