+++ app.js
@@ -369,8 +369,11 @@ const showLine = async (line, socket = io) => {
   }
 };
 
-const updateOverlayVars = () => {
+const updateOverlayVars = overlayVars => {
   const overlayPrefs = store.get('obs');
+  if (overlayVars) {
+    overlayPrefs.overlayVars = { ...overlayVars };
+  }
   io.emit('update-prefs', overlayPrefs);
 };
 
@@ -551,19 +554,9 @@ ipcMain.on('clear-apv', () => {
 let lastLine;
 
 ipcMain.on('update-overlay-vars', updateOverlayVars);
-ipcMain.on('save-overlay-settings', (event, overlayVars) => {
-  store.set('obs', {
-    overlayPrefs: {
-      overlayVars,
-    },
-  });
-  updateOverlayVars();
+ipcMain.on('save-overlay-prefs', (event, overlayVars) => {
+  updateOverlayVars(overlayVars);
 });
-// ipcMain.on('update-overlay-vars', (event, arg) => {
-//   console.log('####### Updated bani overlay vars ##########');
-//   console.log(arg);
-//   io.emit('update-prefs', arg);
-// });
 
 io.on('connection', socket => {
   updateOverlayVars();
